Rename class and add error logging based on status
