+++ app/lib/slack_service/error.rb
@@ -1,0 +1,4 @@
+module SlackService
+  class Error < StandardError
+  end
+end
